Apply scale_factor if one is present
Rather than hard-coding the scale_factor application for SSS and salinity,
marbl_io_read_field checks to see if the netcdf file has a scale_factor
attribute and applies it if it is there.

This required updating netcdf_check to allow the user the option to specify an
error code to ignore (in this case, NF90_ENOTATT signifies the variable does
not have a scale_factor attribute).