Update pylintrc to ignore bad option values
Depending on the environment, pylintrc may throw a useless-object-inheritance
warning in code_consistency, or it may not know what that warning is (so having
an explicit disable throws a bad option value error). Telling pylintrc to
ignore bad option values means that we can disable useless-object-inheritance
regardless of whether the current system will catch that warning or not.