Standalone driver fix for nag
The nag compiler was crashing in the comp_co2calc_coeffs() call from
marbl_co2calc_interior(), and I traced it to the fact that the netCDF
fill values in salinity were leading to the compiler taking the sqrt()
of a negative number. Replacing these fill values with 0 when reading
the data in is a good work around, but the real solution is to have
MARBL only do computations on active levels.

Unfortunately, this does not seem to be related to the problems with
gfortran 4.8.