No longer rely on curly braces in make clean
Found that some versions of Make (or maybe some OSes? or some shells?) don't
like commands such as

$ rm -f $(OBJ_ROOT)/{gnu,pgi}/*.o

If one (or both) of the directories in {} are already empty. So now the
Makefiles pipe the results of find through "xargs rm -f", which is not as
explicit as before but is more reliable.