Move co2calc types onto interface
This is the first step towards moving the dimension (num_elements for surface
co2calc coeffs and state, num_levels for interior) inside the two co2calc
datatypes. The next step will be to add a constructor (and destructor) to the
class, and at that point the dimensions in each subroutine and associate
statement will need to be updated.