push dimension inside co2calc types
co2calc_coeffs_type and co2calc_state_type now contain allocatable elements
rather than declaring arrays of the type with scalar elements. This is a
workaround for a gfortran 4.8 bug -- note that we also removed the
co2calc_state_in local variable and explicitly pass the 6 variables that
collected in the type previously.