Update associate statements due to gfort 4.8 bug
associate statements of the form

associate(new_var => some_type(:)%old_var)

are not interpretted correctly in gfortran 4.8 -- new_var points to a
contiguous block of memory rather than having a stride equal to the size of a
single some_type array element. The previous commit addressed this issue in the
co2calc datatypes, but it was also causing problems in the autotroph_settings
and zooplankton_settings types. In those cases, the workaround was to avoid
using associates altogether.