Move location of test suite temporary file
run_test_suite.sh dumps the results of each run (PASS or FAIL) into a temporary
file and then writes the contents of the file to stdout after running all the
tests so that the user can see a summary of all the results. Originally this
file was in the top MARBL directory, but it makes more sense to have it in
MARBL_tools/ because that is where run_test_suite.sh is. I also renamed it from
.test.out to .test_suite.cache so it's a little clearer what it's for (and
run_test_suite.sh now refers to it as ${RESULTS_CACHE} rather than $OUTFILE)