Major refactor of IO routines
marbl_netcdf_mod.F90 just contains wrappers to nf90_ subroutines (with a clean
abort if the routines are called when MARBL is built without _NETCDF... this
allows tests that do not require netCDF to build and run on machines without
any knowledge of the netCDF library)

marbl_io_mod.F90 is the only module that calls marbl_netcdf_mod.F90. It
contains two ncid variables at the module level (ncid_in for the initial
condition file and ncid_out for the history file), and varids / dimids are no
longer stored at all. Instead, the module makes use of the inq_dimids and
inq_varids subroutines (wrapped in marbl_netcdf_mod.F90).

Did a little clean-up of the netcdf_comparison.py tool so that the variable
comparison tests are still run if the headers of the two files don't match, and
fixed a bug that prevented the driver log from being written in certain failure
modes.