Add calls to set_global_scalars()
The stand-alone test that calls surface_flux_compute() and
interior_tendency_compute() should call set_global_scalars() beforehand. This
isn't strictly necessary for the test (there are no global scalars to set in
the surface flux case, and unless the test is run with ladjust_bury_coeff =
.true. there aren't any in the interior tendency case either), but I want to
include these calls because this test may be used as a blue-print for including
MARBL in new GCMs. All the sphinx documentation will refer to the test instead
of the POP driver for examples, and we definitely want this call documented.