Better comments surrounding update to domain info
For the stand-alone test, each column needs to update kmt on the domain_type.
As an example for GCMs, this step should be explicitly commented so users know
it is also where level depths / thicknesses would need to be updated in models
that do not use the same vertical grid in all cells.