Update settings_dict to hold more data
rather than settings_dict[varname] being set to the value of the variable, it
is now a dictionary with value, units, and longname keys. This is being done to
make it easier to turn the YAML file into a table to publish either in a
journal article or online through sphinx.