Adjust integrated_terms for budget checks
Rather than change the kmt layer of the term we are integrating, I adjust the
kmt level of the terms being summed for the comparison check. I think it looks
a little cleaner, and it avoids dividing by delta_z(kmt) when adding a term to
work (which is then multiplied by delta_z).