Standalone driver: no bottom flux in tendency
For the call_compute_subroutines test, I had been adding bottom_flux / dz to
the tendency to ensure the tendencies were within round-off of the previous
values. That was just for testing purposes, and now that this PR is done with
testing it makes sense to generate fresh baselines where the bottom fluxes and
tendencies are output separately.