Re-order tests in run_test_suite.yml
Setting "continue-on-error: True" for the documentation resulted in the test
passing even though "Build docs" failed -- I was hoping the test would fail,
but that I would also be able to see if the "Run Test Suite" step passed or
failed.

I'd rather be notified of a failure of the python code and not know the status
of the documentation build than the other way around.