interior_tendency_output -> get_output_for_GCM()
MOM6 needs to know the initial total_Chl field, so waiting until we call
interior_tendency_compute() is not sufficient. So rather than getting totalChl
in an analagous manner to how POP gets surface Chl, there is a new function on
the interface [get_output_for_GCM()] that takes the tracer array and an integer
index indicating what field to return as an intent(in).

Note that this new function replaces negative tracer values with 0, but does
not enforce consistency among the autotroph tracers (so Chl may be positive
while C, P, or Fe are 0). Because of this, I expect this commit to fail CI but
I'll update the baseline in the next commit and everything should pass again.