Add support for derecho
Currently allows for intel, cray, nvhpc, and pgi compilers. Note that I don't
support the various permutations of intel (intel-oneapi, intel-classic); I
think intel-classic uses the same fortran compiler as intel, and I don't know
if the fortran compiler in intel-oneapi is still named "ifort" or if it is
something else.