First commit following code review
Keith and I reviewed the fortran changes in src/ and this commit addresses
concerns raised during that review. Most of it is straight forward, I do want
to highlight one change that I think still needs improvement:

Instead of passing unit_system_opt ('cgs' or 'mks') to
marbl_instance%surface_flux_output%add_output(), we call
marbl_instance%get_conc_flux_units and then pass the conc_flux_units directly.
This still feels kludgy, but the way the interface and the
marbl_output_for_GCM_type are defined it was the only option for a quick fix.
I'd like to revisit this interface and come up with a complete redesign of the
way we register fields the GCM would like MARBL to return.