More cleanup for running with base_bio_on = .false.
If base_bio_on = .false. then PFT_defaults is treated as "None" regardless of
its value (a message is added to the log, similar to the way cocco and 4p2z are
treated as user-specified).

Also realized a better solution to the problem addressed in the previous commit
is to move PFT_defaults to general_parms2 so that its default can depend on
base_bio_on; this means that autotroph_cnt and other _cnt variables can be set
to zero when PFT_defaults = None, which is the new default for base_bio_on =
.false.

Lastly, cleaned up the diagnostics yaml file to note which diagnostics should
only be defined when base_bio_on is true