Limit settings vars exposed through put_setting()
The tracer module control variables (base_biotic_on, abio_dic_on, ciso_on) are
always available to the GCM via put_setting(), but if base_biotic_on = .false.
then changing many of the other settings will have no affect on the run. Rather
than giving the false impression that users could set (say) PFT_defaults,
calling put_settings() for variables that will not be used results in an error.

Note that the defined_if key in the settings YAML files matches the logic
around whether or not this%add_var() is called -- so generating a settings file
with MARBL_generate_settings_file.py will always result in a file that is
consistent with the new logic in the Fortran code.

The most noticeable difference from this PR will be seen in the
gen_settings_file test, which will produce a significantly smaller file
settings file using the marbl_with_abio_only.settings input. The POP and MOM6
GCMs will already be using shortened settings files because of the defined_if
key in the YAML.