Add new netcdf_metadata_check.py script
This script compares the units and longname attribute of every variable in a
netCDF history file to what is defined in a diagnostics YAML file. In other
words, it checks for consistency between the Fortran code (which writes the
metadata to netCDF) and the YAML file (which is more human-readable if users
want to look up what diagnostics are available, and which eventually will be
used to auto-generate documentation on the available diagnostics).

Running this test found errors in both the Fortran and the YAML, which have
been fixed. It also uncovered a bug in how the longname of the grazing
variables for zooplankton grazing on zooplankton were parsed.

I fixed a bug where we were defining diagnostics for the base biotic tracers
even when running with base_biotic_on=.false., so the current baseline for the
abio_only test has more variables than the test now generates. I think this
will cause one of the CI tests to fail (and there might be an issue with
parsing the 4p2z settings file causing additional failures); I'll clean those
failures up in the next commit.