Update baselines to account for unit changes
There was a change to the units of Fefree in the last commit, so the netcdf
comparison script could not compare the field to the values in the baseline.
The new baselines for the default test, ciso test, and 4p2z now have proper
units.

Also, the baseline for the abio test only contains the abio fields (the last
commit changed what fields are written when base_biotic_on = .false.) and has
been renamed to explicitly note that it is abio_only rather than both the abio
and base biotic modules.