Mechanism for GCM to disallow vars in input_file
In CESM+MOM6, we want the values of abio_dic_on and base_bio_on to be
determined by the MARBL_TRACER_OPTS variable in env_run.xml; in that case, we
do not want users to set those variables in user_nl_marbl. This commit allows
GCMs to give MARBL a dictionary, where the key is a variable name and the value
is an error message to return if that variable appears in the input_file.